requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-docx
pydantic
mammoth
docx2txt